extracting editModal utils file
