+++ apps/google-docs/src/types/index.ts
@@ -1,4 +1,5 @@
 export * from './entry';
 export * from './entryBlockGraph';
+export * from './editModal';
 export * from './normalizedDocument';
 export * from './workflow';
